+++ comp2013/src/main/java/comp2013/SnakeFood.java
@@ -3,9 +3,7 @@ package comp2013;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
 import javafx.scene.canvas.Canvas;
-import javafx.scene.canvas.GraphicsContext;
 
-import java.awt.*;
 import java.util.Random;
 
 public class SnakeFood
@@ -18,7 +16,7 @@ public class SnakeFood
 	private int M_X, M_Y;
 	public SnakeFood()	{
 		// Gets a random image from the hash map and sets it.
-		this.M_Image = ImageUtil.getImage(String.valueOf(new Random().nextInt(10)));
+		this.M_Image = SnakeImageUtil.getImage(String.valueOf(new Random().nextInt(10)));
 		this.m_Eaten = true;
 		// Get the width and height of the image.
 		this.M_Width = M_Image.getWidth();
@@ -66,7 +64,7 @@ public class SnakeFood
 	 */
 	public void newFruit(){
 		// Carries out the same as the constructor, but can be called by other functions.
-		this.M_Image = ImageUtil.getImage(String.valueOf(new Random().nextInt(10)));
+		this.M_Image = SnakeImageUtil.getImage(String.valueOf(new Random().nextInt(10)));
 		this.m_Eaten = false;
 		this.M_Width = M_Image.getWidth();
 		this.M_Height = M_Image.getHeight();