+++ docs/changelog.md
@@ -106,3 +106,11 @@
 - Updated score funtion so that it now shows the score as food is collected.
 - Updated how the snake self collision is handled, removed an extra check.
 - Added a method to view "gameOverScreen" that handles what happens when a game over event is triggered.
+
+### 9/12/23
+- Fixed issue where boundary detection is to sensitive.
+- Removed Play and MyFrame Classes.
+- Renamed Movable to IMovable.
+- Renamed ImageUtil to SnakeImageUtil
+- Removed SnakeThread Class.
+- Renamed MusicPlayer to SnakeMusic.